Remove python package use case for simplicity (#19)
The main focus of this template should be on the `project template` and
`omniverse extension` usages, rather than `python package`. This change
shall remove the `python package` use case to keep the template as
simple as possible.

Furthermore, the `pip install -e .` instruction set already implies this
template to be a python package.